requirements.txt
pandas
requests
SQLAlchemy==1.4
python-dotenv
psycopg2
utils
skforecast
neptune